Quick and Dirty fixes
Adding colors for the new themes dark theme. Does not apply for any other theme yet. Trying to revert as much as possible. Moving the user on the bottom is sketch still works and abides by the rescaling. Have to use !important a lot for some reason need to investigate further